+++ java/net/lag/jaramiko/sftp/Attributes.java
@@ -27,14 +27,14 @@ public class Attributes implements Comparable<Attributes> {
     public static Attributes fromMessage(Message message) {
         Attributes attributes = new Attributes();
         attributes.flags = message.getInt();
-        if ((attributes.flags & FLAG_SIZE) != 0)
-            attributes.st_size = message.getInt64();
+        if ((attributes.flags & FLAG_SIZE) != 0) attributes.st_size = message
+                .getInt64();
         if ((attributes.flags & FLAG_UIDGID) != 0) {
             attributes.st_uid = message.getInt();
             attributes.st_gid = message.getInt();
         }
-        if ((attributes.flags & FLAG_PERMISSIONS) != 0)
-            attributes.st_mode = message.getInt();
+        if ((attributes.flags & FLAG_PERMISSIONS) != 0) attributes.st_mode = message
+                .getInt();
         if ((attributes.flags & FLAG_AMTIME) != 0) {
             attributes.st_atime = message.getInt();
             attributes.st_mtime = message.getInt();
@@ -50,128 +50,90 @@ public class Attributes implements Comparable<Attributes> {
         return attributes;
     }
 
-
-
     public Long getSt_size() {
         return st_size;
     }
 
-
-
     public void setSt_size(Long st_size) {
         this.st_size = st_size;
-        if (st_size != null)
-            flags |= FLAG_SIZE;
+        if (st_size != null) flags |= FLAG_SIZE;
         else
             flags &= ~FLAG_SIZE;
     }
-
-
 
     public Integer getSt_uid() {
         return st_uid;
     }
 
-
-
     public void setSt_uid(Integer st_uid) {
         this.st_uid = st_uid;
-        if (st_uid != null && this.st_gid != null)
-            flags |= FLAG_UIDGID;
+        if (st_uid != null && this.st_gid != null) flags |= FLAG_UIDGID;
         else
             flags &= ~FLAG_UIDGID;
     }
-
-
 
     public Integer getSt_gid() {
         return st_gid;
     }
 
-
-
     public void setSt_gid(Integer st_gid) {
         this.st_gid = st_gid;
-        if (st_uid != null && this.st_uid != null)
-            flags |= FLAG_UIDGID;
+        if (st_uid != null && this.st_uid != null) flags |= FLAG_UIDGID;
         else
             flags &= ~FLAG_UIDGID;
     }
-
-
 
     public Integer getSt_mode() {
         return st_mode;
     }
 
-
-
     public void setSt_mode(Integer st_mode) {
         this.st_mode = st_mode;
-        if (st_mode != null)
-            flags |= FLAG_PERMISSIONS;
+        if (st_mode != null) flags |= FLAG_PERMISSIONS;
         else
             flags &= ~FLAG_PERMISSIONS;
     }
-
-
 
     public Integer getSt_atime() {
         return st_atime;
     }
 
-
-
     public void setSt_atime(Integer st_atime) {
         this.st_atime = st_atime;
-        if (st_atime != null && this.st_mtime != null)
-            flags |= FLAG_AMTIME;
+        if (st_atime != null && this.st_mtime != null) flags |= FLAG_AMTIME;
         else
             flags &= ~FLAG_AMTIME;
     }
-
-
 
     public Integer getSt_mtime() {
         return st_mtime;
     }
 
-
-
     public void setSt_mtime(Integer st_mtime) {
         this.st_mtime = st_mtime;
-        if (st_atime != null && st_mtime != null)
-            flags |= FLAG_AMTIME;
+        if (st_atime != null && st_mtime != null) flags |= FLAG_AMTIME;
         else
             flags &= ~FLAG_AMTIME;
     }
 
-
     public void toMessage(Message message) {
         this.flags = 0;
-        if (this.st_size != null)
-            this.flags |= FLAG_SIZE;
-        if (this.st_uid != null && this.st_gid != null)
-            this.flags |= FLAG_UIDGID;
-        if (this.st_mode != null)
-            this.flags |= FLAG_PERMISSIONS;
-        if (this.st_atime != null && this.st_mtime != null)
-            this.flags |= FLAG_AMTIME;
-        if (!this.attr.isEmpty())
-            this.flags |= FLAG_EXTENDED;
+        if (this.st_size != null) this.flags |= FLAG_SIZE;
+        if (this.st_uid != null && this.st_gid != null) this.flags |= FLAG_UIDGID;
+        if (this.st_mode != null) this.flags |= FLAG_PERMISSIONS;
+        if (this.st_atime != null && this.st_mtime != null) this.flags |= FLAG_AMTIME;
+        if (!this.attr.isEmpty()) this.flags |= FLAG_EXTENDED;
         message.putInt(this.flags);
-        if ((this.flags & FLAG_SIZE) != 0)
-            message.putInt64(this.st_size);
+        if ((this.flags & FLAG_SIZE) != 0) message.putInt64(this.st_size);
         if ((this.flags & FLAG_UIDGID) != 0) {
             message.putInt(this.st_uid);
             message.putInt(this.st_gid);
         }
-        if ((this.flags & FLAG_PERMISSIONS) != 0)
-            message.putInt(this.st_mode);
+        if ((this.flags & FLAG_PERMISSIONS) != 0) message.putInt(this.st_mode);
         if ((this.flags & FLAG_AMTIME) != 0) {
             // throw away any fractional seconds
-            message.putInt((int) this.st_atime);
-            message.putInt((int) this.st_mtime);
+            message.putInt(this.st_atime);
+            message.putInt(this.st_mtime);
         }
         if ((this.flags & FLAG_EXTENDED) != 0) {
             message.putInt(this.attr.size());
@@ -182,6 +144,7 @@ public class Attributes implements Comparable<Attributes> {
         }
         return;
     }
+
     public String getFilename() {
         return filename;
     }
@@ -244,11 +207,14 @@ public class Attributes implements Comparable<Attributes> {
             if (result == 0) {
                 result = (st_gid == o.st_gid) ? 0 : st_gid.compareTo(o.st_gid);
                 if (result == 0) {
-                    result = (st_mode == o.st_mode) ? 0 : st_mode.compareTo(o.st_mode);
+                    result = (st_mode == o.st_mode) ? 0 : st_mode
+                            .compareTo(o.st_mode);
                     if (result == 0) {
-                        result = (st_atime == o.st_atime) ? 0 : st_atime.compareTo(o.st_atime);
+                        result = (st_atime == o.st_atime) ? 0 : st_atime
+                                .compareTo(o.st_atime);
                         if (result == 0) {
-                            result = (st_mtime == o.st_mtime) ? 0 : st_mtime.compareTo(o.st_mtime);
+                            result = (st_mtime == o.st_mtime) ? 0 : st_mtime
+                                    .compareTo(o.st_mtime);
                         }
                     }
                 }
